Refactor all the things
