+++ googletrans/__init__.py
@@ -1,3 +1,7 @@
 """Free Google Translate API for Python. Translates totally free of charge."""
+__all__ = 'Translator',
+__version_info__ = 2, 0, 0
+__version__ = '.'.join(str(v) for v in __version_info__)
 
-__version__ = 1.2
+
+from googletrans.client import Translator
